Converted statistics from double to big decimal & added spring profile for heroku

+++ src/main/kotlin/com/n26/zehraerguven/dto/TransactionRequest.kt
@@ -1,9 +1,9 @@
 package com.n26.zehraerguven.dto
 
-import com.fasterxml.jackson.annotation.JsonFormat
+import java.math.BigDecimal
 import java.util.*
 
 data class TransactionRequest(
     var timestamp: Date,
-    var amount: Double
+    var amount: BigDecimal
     )
